refactor: Exctract empty method from module
