+++ crates/testa-hir/src/lib.rs
@@ -3,6 +3,9 @@ pub mod module;
 pub mod source_map;
 pub mod string_pool;
 
+#[cfg(feature = "test-utils")]
+mod test_utils;
+
 pub use lowering::AstLowering;
 pub use module::{FieldId, Item, ItemId, Module, ModuleMetadata, StringId};
 pub use string_pool::StringPool;
