+++ crates/testa-hir/src/module/mod.rs
@@ -306,23 +306,6 @@ pub struct ModuleMetadata {
 }
 
 impl Module {
-    pub fn empty(name: &str) -> Self {
-        Self {
-            metadata: ModuleMetadata {
-                version: 1,
-                name: name.to_string(),
-                source_file: PathBuf::from(format!("{}.testa", name)),
-                source_hash: 0,
-                compiled_at: 0,
-            },
-            string_pool: StringPool::new(),
-            imports: Vec::new(),
-            items: Vec::new(),
-            directives: Vec::new(),
-            source_map: SourceMap::new(Vec::new(), Vec::new(), Vec::new()),
-        }
-    }
-
     pub fn get_item(&self, id: ItemId) -> Option<&Item> {
         self.items.iter().find(|item| match item {
             Item::Template(t) => t.id == id,
