+++ crates/testa-interpreter/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2024"
 
 [dependencies]
 testa-core = { path = "../testa-core" }
-testa-hir = { path = "../testa-hir" }
+testa-hir = { path = "../testa-hir", features = ["test-utils"] }
 rand = "0.9.2"
 once_cell = "1.21.3"
 indexmap = { version = "2.12.1", features = ["serde"] }
